Simplified the consumer code
